requirements.txt
# FastMCP for MCP server
fastmcp>=2.12.5

# OpenAI for content generation
openai>=1.0.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# HTML parsing and sanitization
beautifulsoup4>=4.12.0
html5lib>=1.1

